Feat: Move ScannedBikesViewModel to its own file
